# PAW-API_Producao/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Informação da RestAPI Produção",
        "version": "1.0.0",
        "title": "PAW - RestAPI Produção",
        "contact": {
            "name": "Grupo - Carlos Pinto, João Pires, Nuno Silva, Sérgio Rodrigues"
        }
    },
    "schemes": [
        "http"
    ],
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/productionorders": {
            "get": {
                "summary": "Obter todas as ordens de produção",
                "description": "Obter todas as ordens de produção",
                "produces": [
                    "application/json"
                ],
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "operação com sucesso",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ProductionOrders"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            },
            "post": {
                "summary": "Criar nova ordem de produção",
                "description": "Criar nova ordem de produção",
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "ordem de produção",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "type": "object",
                                "$ref": "#/definitions/ProductionOrders"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ProductionOrders"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/productionorders/productionorder?page={nPaginas}&limit={limite}": {
            "get": {
                "summary": "Obter todas as ordens de produção, ordenadas por status e definindo nº de páginas e limite",
                "description": "Obter todas as ordens de produção, ordenadas por status e definindo nº de páginas e limite",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "nPaginas",
                        "description": "Número de páginas"
                    },
                    {
                        "in": "path",
                        "name": "limite",
                        "description": "Limite de ordens de produção a apresentar"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ProductionOrders"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/productionorders/{productionorderId}": { 
            "get": {
                "summary": "Obter ordem de produção por id",
                "description": "Obter ordem de produção por id",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "productionorderId",
                        "description": "id da ordem de produção a pesquisar",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "task": {
                                    "type": "object",
                                    "$ref": "#/definitions/ProductionOrders"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ProductionOrders"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            },
            "patch": {
                "summary": "Atualizar ordem de produção",
                "description": "Atualizar ordem de produção",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "productionorderId",
                        "in": "path",
                        "description": "id da ordem de produção a alterar",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "ordem de produção a alterar",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "type": "object",
                                "$ref": "#/definitions/ProductionOrders"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ProductionOrders"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            },
            "delete": {
                "summary": "Apagar ordem de produção",
                "description": "Apagar ordem de produção",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "productionorderId",
                        "in": "path",
                        "description": "id da ordem de produção a eliminar",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ProductionOrders"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
...
        "/search/productionorders?tags={querie}": {
            "get": {
                "summary": "Obter ordens de produção através de pesquisa",
                "description": "Obter ordens de produção através de pesquisa",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "querie",
                        "description": "termo a pesquisar",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "task": {
                                    "type": "object",
                                    "$ref": "#/definitions/ProductionOrders"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ProductionOrders"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ProductionOrders": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                },
                "demand_quant": {
                    "type": "integer"
                },
                "efect_quant": {
                    "type": "integer"
                },
                "date_created": {
                    "type": "string"
                },
                "date_start": {
                    "type": "string"
                },
                "date_finish": {
                    "type": "string"
                },
                "assigned_to": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "item_id": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "InvalidResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}